Added support for wireless xbox 360 controllers on Ouya
I noticed wireless xbox 360 controllers connected with the wireless
receiver were not working properly on Ouya, so I wrote a device profile.

Unfotunately there seems to be a conflict between the dpad and face
buttons returning the same button mappings, which I've noted with
comments next to the offending buttons.